requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

# Satellite Image Processing
opencv-python==4.8.1.78
Pillow==10.0.1
rasterio==1.3.8
geopandas==0.14.1
shapely==2.0.2
fiona==1.9.5

# Machine Learning & Deep Learning
tensorflow==2.13.0
scikit-learn==1.3.0
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2

# Data Processing & Visualization
plotly==5.15.0
folium==0.14.0
geopy==2.3.0
requests==2.31.0

# File Handling & Utilities
python-dotenv==1.0.0
gunicorn==21.2.0
